refactor: decompose PianoRollCanvas.tsx into focused modules
Extract pure canvas rendering into PianoRollRenderer.ts (388 lines) and
drag state machine into usePianoRollDrag.ts (565 lines), reducing the
main component from 1391 to 784 lines.

Closes #1028
